[Feat #6]: Refacotor project structure
