+++ src/containers/Main/Pokemons/PokemonItem.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { getPokemonId, pokemonImageUrl, formatName } from '../../helpers/utils';
+import { getPokemonId, pokemonImageUrl, formatName } from '../../../helpers/utils';
 
 const PokemonItem = ({ pokemon }) => {
   const id = getPokemonId(pokemon.url);
